--- a/components/box-emu-hal/src/statistics.cpp
+++ b/components/box-emu-hal/src/statistics.cpp
@@ -55,5 +55,5 @@ void print_statistics() {
     fmt::print("-----------\n");
     fmt::print("Frames: {}\n", num_frames);
     fmt::print("FPS: {:.1f}\n", get_fps());
-    fmt::print("Frame Time: [{}, {:.1f}, {}]\n", get_frame_time_min(), get_frame_time_avg(), get_frame_time_max());
+    fmt::print("Frame Time: [min {} us, avg: {:.1f} us, max: {} us]\n", get_frame_time_min(), get_frame_time_avg(), get_frame_time_max());
 }