--- a/components/box-emu-hal/src/video_task.cpp
+++ b/components/box-emu-hal/src/video_task.cpp
@@ -125,6 +125,8 @@ static bool video_task(std::mutex &m, std::condition_variable& cv) {
     // if we don't have a custom palette, we just need to scale/fill the frame
     float y_scale = (float)display_height/native_height;
     float x_scale = (float)display_width/native_width;
+    float inv_x_scale = (float)native_width/display_width;
+    float inv_y_scale = (float)native_height/display_height;
     int max_y = lcd_height;
     int max_x = std::clamp<int>(x_scale * native_width, 0, lcd_width);
     for (int y=0; y<max_y; y+=num_lines_to_write) {
@@ -138,14 +140,14 @@ static bool video_task(std::mutex &m, std::condition_variable& cv) {
         if (_y >= max_y) {
           break;
         }
-        int source_y = (float)_y/y_scale;
+        int source_y = (float)_y * inv_y_scale;
         // shoudl i put this around the outer loop or is this loop a good
         // balance for perfomance of the check?
         if (has_palette()) {
           const uint8_t* _frame = (const uint8_t*)_frame_ptr;
           // TODO: write two pixels (32 bits) at a time because it's faster
           for (int x=0; x<max_x; x++) {
-            int source_x = (float)x/x_scale;
+            int source_x = (float)x * inv_x_scale;
             int index = source_y*native_pitch + source_x;
             _buf[i*max_x + x] = _palette[_frame[index] % palette_size];
           }
@@ -153,7 +155,7 @@ static bool video_task(std::mutex &m, std::condition_variable& cv) {
           const uint16_t* _frame = (const uint16_t*)_frame_ptr;
           // TODO: write two pixels (32 bits) at a time because it's faster
           for (int x=0; x<max_x; x++) {
-            int source_x = (float)x/x_scale;
+            int source_x = (float)x * inv_x_scale;
             _buf[i*max_x + x] = _frame[source_y*native_pitch + source_x];
           }
         }