--- a/components/gbc/src/gameboy.cpp
+++ b/components/gbc/src/gameboy.cpp
@@ -155,7 +155,7 @@ void init_gameboy(const std::string& rom_filename, uint8_t *romdata, size_t rom_
 }
 
 void run_gameboy_rom() {
-  auto start = std::chrono::steady_clock::now();
+  auto start = esp_timer_get_time();
   // GET INPUT
   InputState state;
   hal::get_input_state(&state);
@@ -176,13 +176,13 @@ void run_gameboy_rom() {
   }
   last_x = state.x;
 
-  auto end = std::chrono::steady_clock::now();
-  auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(end-start);
-  auto elapsed_float = std::chrono::duration<float>(elapsed).count();
-  auto max_frame_time = std::chrono::milliseconds(15);
-  update_frame_time(elapsed_float);
+  auto end = esp_timer_get_time();
+  auto elapsed = end - start;
+  update_frame_time(elapsed);
+  static constexpr uint64_t max_frame_time = 1000000 / 60;
   if (!unlock && elapsed < max_frame_time) {
-    std::this_thread::sleep_for(max_frame_time - elapsed);
+    auto sleep_time = (max_frame_time - elapsed) / 1e3;
+    std::this_thread::sleep_for(sleep_time * 1ms);
   }
 }