--- a/components/genesis/CMakeLists.txt
+++ b/components/genesis/CMakeLists.txt
@@ -5,5 +5,5 @@ idf_component_register(
   REQUIRES box-emu-hal
   )
 # target_compile_options(${COMPONENT_LIB} PRIVATE -Wno-char-subscripts -Wno-attributes -Wno-implicit-fallthrough -Wno-unused-function -Wno-unused-variable -Wno-discarded-qualifiers)
-target_compile_options(${COMPONENT_LIB} PRIVATE -Wno-unused-const-variable -O3)
+target_compile_options(${COMPONENT_LIB} PRIVATE -Wno-unused-const-variable -Wno-unused-value -O3)
 # target_compile_definitions(${COMPONENT_LIB} PRIVATE GWENESIS_AUDIO_ACCURATE=0)