--- a/components/genesis/gwenesis/src/cpus/M68K/m68kcpu.c
+++ b/components/genesis/gwenesis/src/cpus/M68K/m68kcpu.c
@@ -21,6 +21,8 @@ extern int vdp_68k_irq_ack(int int_level);
   #endif
 #endif
 
+#include <esp_attr.h>
+
 #include "m68kconf.h"
 #include "m68kcpu.h"
 #include "m68kops.h"
@@ -261,7 +263,7 @@ void m68k_set_irq_delay(unsigned int int_level)
   m68ki_check_interrupts(); /* Level triggered (IRQ) */
 }
 
-void m68k_run(unsigned int cycles) 
+void IRAM_ATTR m68k_run(unsigned int cycles)
 {
     //  printf("m68K_run current_cycles=%d add=%d STOP=%x\n",m68k.cycles,cycles,CPU_STOPPED);