--- a/components/sms/src/sms.cpp
+++ b/components/sms/src/sms.cpp
@@ -105,7 +105,7 @@ void init_gg(uint8_t *romdata, size_t rom_data_size) {
 }
 
 void run_sms_rom() {
-  auto start = std::chrono::high_resolution_clock::now();
+  auto start = esp_timer_get_time();
   // handle input here (see system.h and use input.pad and input.system)
   InputState state;
   hal::get_input_state(&state);
@@ -176,15 +176,9 @@ void run_sms_rom() {
   hal::play_audio((uint8_t*)sms_audio_buffer, sms_audio_buffer_len * 2 * 2); // 2 channels, 2 bytes per sample
 
   // manage statistics
-  auto end = std::chrono::high_resolution_clock::now();
-  auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(end-start);
-  auto elapsed_float = std::chrono::duration<float>(elapsed).count();
-  update_frame_time(elapsed_float);
-  // NOTE: seems like it doesn't need this...
-  // using namespace std::chrono_literals;
-  // if (elapsed < 15ms) {
-  //   std::this_thread::sleep_for(15ms - elapsed);
-  // }
+  auto end = esp_timer_get_time();
+  auto elapsed = end - start;
+  update_frame_time(elapsed);
 }
 
 void load_sms(std::string_view save_path) {