--- a/components/sms/src/sms.cpp
+++ b/components/sms/src/sms.cpp
@@ -30,6 +30,8 @@ static int frame_buffer_index = 0;
 
 static uint32_t *sms_audio_buffer = nullptr;
 
+static bool unlock = false;
+
 void sms_frame(int skip_render);
 void sms_init(void);
 void sms_reset(void);
@@ -73,6 +75,9 @@ static void init(uint8_t *romdata, size_t rom_data_size) {
 
   frame_counter = 0;
   muteFrameCount = 0;
+
+  // reset unlock
+  unlock = false;
 
   initialized = true;
   reset_frame_time();
@@ -175,10 +180,22 @@ void run_sms_rom() {
   // push the audio buffer to the audio task
   hal::play_audio((uint8_t*)sms_audio_buffer, sms_audio_buffer_len * 2 * 2); // 2 channels, 2 bytes per sample
 
+  // update unlock based on x button
+  static bool last_x = false;
+  if (state.x && !last_x) {
+    unlock = !unlock;
+  }
+  last_x = state.x;
+
   // manage statistics
   auto end = esp_timer_get_time();
   auto elapsed = end - start;
   update_frame_time(elapsed);
+  static constexpr uint64_t max_frame_time = 1000000 / 60;
+  if (!unlock && elapsed < max_frame_time) {
+    auto sleep_time = (max_frame_time - elapsed) / 1e3;
+    std::this_thread::sleep_for(sleep_time * std::chrono::milliseconds(1));
+  }
 }
 
 void load_sms(std::string_view save_path) {