<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
* Switch frame timing to use `esp_timer_get_time()` for more precision and less overhead
* Update video task to perform the display scaling divide and store it as an inverse factor for multiplication in the loops for faster speed
* Update video task so the loops iterate over half as many values, writing two values (32 bits) at a time for slightly faster speed
* Allow nes, gameboy/color, sms emulators to unlock their frame timing (toggled with the `x` button), otherwise they target 60 Hz (sleeping to wait)
* Work on improving speed of genesis emulator, moving more of its code to IRAM.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
Overall we want better performance (so we can unlock the emulators and run them faster than real time - because I'm impatient :P) and we need to squeeze more performance out of every subsystem to try to get the genesis emulator to run full speed when audio is enabled (and ideally to not have to skip frame rendering...).

## How has this been tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
Building and running `main` on `box-3-emu`. Testing all the emulators (and the unlocked variants of `nes`, `gbc`, and `sms` emulators.

* Main branch test on Sonic 1: 53.8 FPS average (when frameskip=3, audio on)
* This branch test on Sonic 1: 74.8 FPS average (when frameskip=3, audio on)

  ```
  Frames: 1859
  FPS: 74.8
  Frame Time: [min 5151 us, avg: 13369.2 us, max: 33994 us]
  ```

There's still quite a few frames which exceed the target of 16ms (60Hz), but it's definitely an improvement.

## Screenshots (if appropriate, e.g. schematic, board, console logs, lab pictures):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update
- [ ] Hardware (schematic, board, system design) change
- [x] Software change

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My change requires a change to the documentation.
- [ ] I have added / updated the documentation related to this change via either README or WIKI

### Software
<!-- Delete this section if not relevant to your PR  -->
- [ ] I have added tests to cover my changes.
- [ ] I have updated the `.github/workflows/build.yml` file to add my new test to the automated cloud build github action.
- [x] All new and existing tests passed.
- [x] My code follows the code style of this project.